Ajouts req all 405

+++ trouve_ta_voie_serveur/models/lieuModel.js
@@ -22,12 +22,12 @@ module.exports = (sequelize, Sequelize) => {
                 len: [3, 500]
             }
         },
-        // edit: Verifié si les co sont valide dans la bd genre qu'il manque pas de chiffre
+        // edit: Google map est encore plus precis et c'est valeur sont arrondis
         latitude: {
-            type: DataTypes.FLOAT(4, 18)
+            type: DataTypes.DECIMAL(18, 15)
         },
         longitude: {
-            type: DataTypes.FLOAT(4, 18)
+            type: DataTypes.DECIMAL(19, 15)
         }
     });
 };
